# docs/swagger.yaml
basePath: /
definitions:
  models.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  models.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  models.Transaction:
    properties:
      amount:
        type: number
      from:
        type: string
      id:
        type: integer
      to:
        type: string
    type: object
  models.Wallet:
    properties:
      address:
        type: string
      balance:
        type: number
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для управления транзакциями и кошельками
  title: Payment System API
  version: "1.0"
paths:
  /api/send:
    post:
      consumes:
      - application/json
      description: Переводит деньги с одного кошелька на другой
      parameters:
      - description: Данные транзакции
        in: body
        name: transaction
        required: true
        schema:
          $ref: '#/definitions/models.Transaction'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Отправить деньги
  /api/transactions:
    get:
      description: Возвращает N последних по времени переводов средств
      parameters:
      - description: Количество транзакций (по умолчанию 10)
        in: query
        name: count
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить последние транзакции
  /api/wallet/{address}/balance:
    get:
      description: Возвращает баланс по адресу кошелька
      parameters:
      - description: Адрес кошелька
        in: path
        name: address
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Wallet'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Получить баланс кошелька
swagger: "2.0"